+++ parranderos/src/main/java/uniandes/edu/co/parranderos/repositorio/SucursalRepository.java
@@ -17,20 +17,17 @@ public interface SucursalRepository extends JpaRepository<Sucursal, Long> {
     @Query(value = "SELECT * FROM SUCURSAL WHERE IDSUCURSAL = :id", nativeQuery = true)
     Sucursal obtenerSucursalPorId(@Param("id") Long id);
 
-    // Insertar una nueva sucursal con relación a la ciudad
     @Modifying
     @Transactional
     @Query(value = "INSERT INTO SUCURSAL (IDSUCURSAL, NOMBRESUCURSAL, TAMANIOINSTALACION, DIRECCION, TELEFONO, CODIGOCIUDAD) " +
-                   "VALUES (:idSucursal, :nombreSucursal, :tamanioInstalacion, :direccion, :telefono, :codigoCiudad)", nativeQuery = true)
+           "VALUES (:idSucursal, :nombreSucursal, :tamanioInstalacion, :direccion, :telefono, :codigoCiudad)", nativeQuery = true)
     void insertarSucursal(@Param("idSucursal") Long idSucursal,
                           @Param("nombreSucursal") String nombreSucursal,
                           @Param("tamanioInstalacion") String tamanioInstalacion,
                           @Param("direccion") String direccion,
                           @Param("telefono") String telefono,
-                          @Param("codigoCiudad") Long codigoCiudad);
+                          @Param("codigoCiudad") Integer codigoCiudad);
     
-
-    // Actualizar una sucursal existente
     @Modifying
     @Transactional
     @Query(value = "UPDATE SUCURSAL SET NOMBRESUCURSAL = :nombreSucursal, TAMANIOINSTALACION = :tamanioInstalacion, " +
@@ -40,7 +37,7 @@ public interface SucursalRepository extends JpaRepository<Sucursal, Long> {
                             @Param("tamanioInstalacion") String tamanioInstalacion,
                             @Param("direccion") String direccion,
                             @Param("telefono") String telefono,
-                            @Param("codigoCiudad") Long codigoCiudad);
+                            @Param("codigoCiudad") Integer codigoCiudad);
     @Query(value = "SELECT s.* FROM SUCURSAL s " +
                             "INNER JOIN BODEGA b ON s.IDSUCURSAL = b.IDSUCURSAL " +
                             "INNER JOIN INFOEXTRABODEGA ie ON b.IDBODEGA = ie.IDBODEGA " +
